git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/branches/KinodynamicRRT@4755 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/DistanceMetrics/RMSDDistance.h
@@ -58,7 +58,7 @@ GetCoordinatesForRMSD(const CfgType& _c) {
   Environment* env = this->GetMPProblem()->GetEnvironment();
   _c.ConfigEnvironment();
   vector<Vector3d> coordinates;
-  for(int i=0; i< env->GetRobot(_c.GetRobotIndex())->NumFreeBody(); ++i)
+  for(size_t i=0; i< env->GetRobot(_c.GetRobotIndex())->NumFreeBody(); ++i)
     coordinates.push_back(env->GetRobot(_c.GetRobotIndex())->GetFreeBody(i)
         ->WorldTransformation().translation());
   return coordinates;
